% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{bioheatmap-shiny}
\alias{bioheatmap-shiny}
\alias{bioheatmapOutput}
\alias{renderBioheatmap}
\title{Shiny bindings for bioheatmap}
\usage{
bioheatmapOutput(output_id, width = "100\%", height = "480px")

renderBioheatmap(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{output_id}{Output variable to read from.}

\item{width, height}{Element size, passed to
\code{\link[htmlwidgets:shinyWidgetOutput]{htmlwidgets::shinyWidgetOutput()}}.}

\item{expr}{An expression that generates a \code{\link[=bioheatmap]{bioheatmap()}} widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} already quoted? Defaults to \code{FALSE}.}
}
\value{
\code{bioheatmapOutput()} returns a Shiny output UI element;
\code{renderBioheatmap()} returns a Shiny render function.
}
\description{
Output and render functions for using \code{\link[=bioheatmap]{bioheatmap()}} within Shiny
applications and interactive R Markdown / Quarto documents.
}
