% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioviz-package.R
\docType{package}
\name{bioviz-package}
\alias{bioviz}
\alias{bioviz-package}
\title{bioviz: High-performance bioinformatics visualizations}
\description{
A collection of GPU-accelerated visualization widgets for large biological
datasets, backed by a shared JavaScript core. Each widget follows the same
pattern: an exported constructor that packs a data frame plus options into a
payload consumed by the bundled JS component.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/samuelbharti/visualization-components}
  \item Report bugs at \url{https://github.com/samuelbharti/visualization-components/issues}
}

}
\author{
\strong{Maintainer}: Samuel Bharti \email{samuel.bharti@posit.co}

Authors:
\itemize{
  \item Samuel Bharti \email{samuel.bharti@posit.co}
}

}
\keyword{internal}
