% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{box_plot_facetted}
\alias{box_plot_facetted}
\title{Function to create boxplots of a metric variable, grouped by two nominal variables}
\usage{
box_plot_facetted(data, group_by1, group_by2, col.jitter_by = NULL,
  shape_jitter_by = NULL, title = NULL, xlab = NULL, ylab = NULL,
  legend = FALSE, rotate = FALSE)
}
\arguments{
\item{data}{A data frame with group variables and a column with numeric values. (Required)}

\item{group_by1}{First grouping variable for which the plot will be facetted by. (Required)}

\item{group_by2}{Second grouping variable which defines the subgroup of group1 (Required)}

\item{col.jitter_by}{The variable name by which the jitter points will be colored by. (Default: NULL)}

\item{shape_jitter_by}{The variable name by which the jitter points will be shaped by. (Default: NULL)}

\item{title}{The title of the plot (Default: NULL)}

\item{xlab}{The x-axis label (Default: NULL)}

\item{ylab}{The y-axis label (Default: NULL)}

\item{legend}{Flag indicating whether the legend should be displayed (Default: FALSE)}

\item{rotate}{Flag indicating whether the x-axis labels should be rotated by 45° (Default: FALSE)}
}
\value{
A ggplot figure
}
\description{
Function to create boxplots of a metric variable, grouped by two nominal variables
}
\examples{
library(reshape2)
n <- 500
df <- data.frame(A=sample(c("a1", "a2", "a3"), n, replace=T),
                 B=sample(c("b1", "b2"), n, replace=T),
                 C=sample(c("c1", "c2", "c3"), n, replace=T),
                 D=sample(c("d1", "d2", "d3"), n, replace=T),
                 value=log(rnorm(n, 100, 1)))

box_plot_facetted(df, group_by1 = "A", group_by2 = "B",
                  col.jitter_by = "A",
                  shape_jitter_by = "C", rotate=T)
}
\author{
Jens Hooge
}

