% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering.hclust_plot}
\alias{clustering.hclust_plot}
\title{Function to plot a heatmap with hirarchical clustering}
\usage{
clustering.hclust_plot()
}
\value{
Not implemented yet!
}
\description{
Function to plot a heatmap with hirarchical clustering
}
