% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{dimred.plot_mca}
\alias{dimred.plot_mca}
\title{Function to plot components from MCA (Multiple Correspondence Analysis)}
\usage{
dimred.plot_mca()
}
\value{
Not implemented yet!
}
\description{
Function to plot components from MCA (Multiple Correspondence Analysis)
}

