% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{dimred.plot_pls}
\alias{dimred.plot_pls}
\title{Function to plot components from PLS (Partial Least Squares)}
\usage{
dimred.plot_pls()
}
\value{
Not implemented yet!
}
\description{
Function to plot components from PLS (Partial Least Squares)
}
