% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{general.bar_plot_simple}
\alias{general.bar_plot_simple}
\title{Function to create a simple barplot}
\usage{
general.bar_plot_simple(freq, labels = NULL, labels.ab = NULL, file.name,
  col = "tomato3", cex.lab = 1, cex.lab.ab = 1, cex.x = 1, xlab = "",
  ylab = "Frequency", add.legend = FALSE, leg.lab = NULL,
  leg.col = NULL, leg.title = NULL, mar = c(5, 4, 3, 1), png.width = 6,
  png.height = 5)
}
\arguments{
\item{freq}{vector with the bar heights, e.g. a frequency table ('table()' output),
names will be used for bar labels if 'labels' is not specified}

\item{labels}{bar labels (default=NULL)}

\item{labels.ab}{additional labels, which are placed above the bars (default=NULL)}

\item{file.name}{name for the PNG output file (without '.png' ending)}

\item{col}{color for the bars (default='tomato3'). if more than one color is specified,}

\item{cex.lab}{size of the bar labels (default=1)}

\item{cex.lab.ab}{size of the bar labels above the bars (default=1)}

\item{cex.x}{size of the x-labels (default=1)}

\item{xlab}{label for the x-axis (default='')}

\item{ylab}{label for the y-axis (default='Frequency')}

\item{add.legend}{TRUE/FALSE should a legend be added underneath the plot 
only works if legend parameters are specified (default=TRUE)}

\item{leg.lab}{labels for the legend (default=NULL)}

\item{leg.col}{colors for the legend (default=NULL)}

\item{leg.title}{title for the legend (default=NULL)}

\item{mar}{plot margin (default=c(5,4,3,1))}

\item{png.width}{width of the PNG-file in inches (default=6)}

\item{png.height}{height of the PNG-file in inches (default=5)}
}
\value{
PNG file with the barplot ('file.name.png') and a 
    numeric vector with size of the PNG-file in inches for rescaling in RTF
}
\description{
Function to create a simple barplot
}
\examples{
\dontrun{
set.seed(42)
x <- sample(1:30, size=500, replace=TRUE)
freq <- table(x)

general.bar_plot_simple(freq, labels.ab=1:30, file.name="test", 
                col=rep(c('tomato3','tomato3','skyblue2'), 10), 
                cex.lab=.7, cex.lab.ab=.8, cex.x=.9,
                ylab="Bar Height", 
                add.legend=TRUE, leg.title="Color",
                leg.lab=c('red','blue'), leg.col=c('tomato3','skyblue2'),
                mar=c(4,4,3,1))
}
}
\author{
Sebastian Voss, Adam Skubala
}

