% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{general.hist_log}
\alias{general.hist_log}
\title{Function to create histograms to check for the necessity of a log-transformation of metric data.}
\usage{
general.hist_log(x, by = NULL, file.name, nbars = 15, title = "",
  xlab = "", col.orig = "lightblue", col.log = "lightyellow")
}
\arguments{
\item{x}{Numeric vector (Required)}

\item{by}{Character or factor vector of the same length than 'x' to 
the time, if two time points should be analyzed. should
contain only two levels (Default: Null)}

\item{file.name}{begining of the name for the PNG output files (without '.png' ending) (Required)}

\item{nbars}{number of bars for the histograms (Default: 15)}

\item{title}{Title of the plot (Default: '')}

\item{xlab}{Label for the x-axis (Default: '')}

\item{col.orig}{Color of the histogram bars of the original values (Default: 'lightblue')}

\item{col.log}{Color of the histogram bars of the log-transformed values (Default: 'lightyellow')}
}
\value{
A .png file with two or four histograms and a numeric vector
    with the size of the .png
}
\description{
The function creates two histograms, one of the original data
and one of the log-tramsformed data. If desired, a second time point can be added,
so that a plot with four histograms is created.
}
\examples{
\dontrun{
x <- exp(rnorm(1000))
by <- factor(rep(c('baseline','day 30'), each=500), 
             level=c('baseline','day 30'))
general.hist_log(x=x, by=by, file.name='test_hist',
         title='Test Historgram', xlab='Measurement')
}
}
\author{
Sebastian Voss, Adam Skubala
}
