% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{general.lineplot}
\alias{general.lineplot}
\title{Function generate (non-)facetted lineplot}
\usage{
general.lineplot(data, by = NULL, facet.rows = 1, upper.err, lower.err,
  var.type = c("errorbar", "ribbon"), title = NULL, xlab = NULL,
  ylab = NULL, xlim = NULL, ylim = NULL, legend.position = "right",
  size = 3, x.log10trans = FALSE, y.log10trans = FALSE)
}
\arguments{
\item{data}{A data frame in long format (Required)}

\item{by}{Variable the plot should be facetted by (Default: NULL)}

\item{facet.rows}{Number of rows in which the facets should be displayed. (Default: 1)}

\item{upper.err}{Upper bound width on errorbars or ribbon variability region (Default: 1)}

\item{lower.err}{Lower bound width on errorbars or ribbon variability region (Default: 1)}

\item{var.type}{Type of how to display variability ("errorbar", "ribbon") (Default: "errorbar")}

\item{title}{Title of the plot (Default: NULL)}

\item{xlab}{x-axis label (Default: NULL)}

\item{ylab}{y-axis label (Default: NULL)}

\item{xlim}{Upper and lower limit (e.g. c(0, 10)) of x-axis (Default: NULL)}

\item{ylim}{Upper and lower limit (e.g. c(0, 10)) of y-axis (Default: NULL)}

\item{legend.position}{Position of the legend ("none", "left", "right", "top", "bottom") (Default: "left")}

\item{size}{Point Size (Default: 3)}

\item{x.log10trans}{Logical flag for log10 x-axis trasformation}

\item{y.log10trans}{Logical flag for log10 y-axis trasformation}
}
\value{
a ggplot2 plot
}
\description{
Function generate (non-)facetted lineplot
}
\examples{
\dontrun{
library(reshape2)
library(ggplot2)
n=7
df <- data.frame(Dose1=exp((1/2)*1:n),
                 Dose2=exp((1/3)*1:n),
                 Dose3=exp((1/4)*1:n),
                 Dose4=exp((1/5)*1:n),
                 Dose5=exp((1/6)*1:n),
                 Dose6=exp((1/7)*1:n),
                 Visit=1:n)
df <- melt(df, id=c("Visit"))
colnames(df) <- c("Visit", "Dose", "Measure") 

## Lineplot with Errorbars
general.lineplot(df, upper.err=1, lower.err=1, var.type="errorbar")

## Lineplot with Errorbars and log transformed y axis
general.lineplot(df, upper.err=1, lower.err=1, var.type="errorbar", y.log10trans = T)

## Facetted Lineplot
general.lineplot(df, by="Dose", facet.row=3, upper.err=1, lower.err=1, 
                 var.type="errorbar", y.log10trans = T)
                 
## Lineplot with Error Ribbon 
general.lineplot(df, by="Dose", upper.err=1, lower.err=1, var.type="ribbon")

## Lineplot with Error Ribbon and log transformed y-axis
general.lineplot(df, by="Dose", upper.err=1, lower.err=1, var.type="ribbon", y.log10trans = T)
}
}
\author{
Jens Hooge
}
