% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{general.scatter.facetted}
\alias{general.scatter.facetted}
\title{Function to generate a facetted scatter plot}
\usage{
general.scatter.facetted(data, by, fun = "mean", smooth.fun = "auto",
  smooth.formula = y ~ x, title = NULL, xlab = NULL, ylab = NULL,
  legend.pos = "none")
}
\arguments{
\item{data}{A dataframe with at least three columns (x,y and a facetting variable) (Required)}

\item{by}{Grouping variable (Required)}

\item{fun}{Function for statistical aggregation function (Default: "mean")}

\item{smooth.fun}{Smoothing method (function) to use, eg. "lm", "glm", "gam", "loess", "rlm". (Default: "auto")
For method = "auto" the smoothing method is chosen based on the size of the largest group 
(across all panels). loess is used for less than 1,000 observations; otherwise gam is used with 
formula = y ~ s(x, bs = "cs"). Somewhat anecdotally, loess gives a better appearance, 
but is O(n^2) in memory, so does not work for larger datasets.}

\item{smooth.formula}{Formula to use in smoothing function, eg. y ~ x, y ~ poly(x, 2), y ~ log(x) (Default: y ~ x)}

\item{title}{Plot Title (Default: NULL)}

\item{xlab}{x-axis label (Default: NULL)}

\item{ylab}{y-axis label (Default: NULL)}

\item{legend.pos}{(Default: "none")}
}
\value{
ggplot2 figure
}
\description{
Function to generate a facetted scatter plot
}
\author{
Jens Hooge
}
\keyword{internal}

