% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{general.volcano_plot}
\alias{general.volcano_plot}
\title{Plots a volcano plot.}
\usage{
general.volcano_plot(data, title = "Volcano Plot", groups = c("Untreated",
  "Treated"), top_labeled = 0, xlabel = expression(log[2]("Foldchange")),
  ylabel = expression(-log[10]("p-value")), xcutoff = c(-log2(4), log2(4)),
  ycutoff = -log10(0.05), log1p = FALSE)
}
\arguments{
\item{data}{A data.frame which includes columns for -log10(p-values) and log2(foldchange)}

\item{title}{The figure title}

\item{groups}{A character vector with the treatment groups}

\item{top_labeled}{A numeric value that controls how many samples should be highlighted based on their pvalue}

\item{xlabel}{X-Axis Label}

\item{ylabel}{Y-Axis Label}

\item{xcutoff}{A numeric value for the -log10(p-value) cutoff}

\item{ycutoff}{A numeric value for the log2(foldchange) cutoff}

\item{log1p}{A flag to apply a log(1+x) transformation on the y axis to account for large values}
}
\value{
A ggplot2 figure
}
\description{
The plot does not aim to show any details about which genes or
pathways are over/under represented, but is meant to show only
the overall directionality of the genes between groups. If there
is a certain perturbation in a system, the volcano plot may reveal
a response in which there is a dramatic increase or decreased in
transcription. In either case, we can get a better idea of the overall
system by showing a volcano plot. Each gene is represented with its p-value
as the x-axis and its (log2) fold change as the y-axis.
}
\examples{
\dontrun{
data(iPSC)
## Plot Simple Volcano Plot
general.volcano_plot(iPSC)

## Plot Simple Volcano Plot with Group Labels
general.volcano_plot(iPSC, groups=c("A", "B"))

## Plot Volcano Plot and add top10 most significant samples
general.volcano_plot(iPSC, top_labeled=10, xcutoff=c(-log2(2), log2(4)))

## Plot Volcano Plot and add Top10 most Significant Samples and Transform y-axis by log(1+x)
general.volcano_plot(iPSC, top_labeled=10, log1p=TRUE)
}
}
\author{
Jens Hooge
}
