% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosling.R
\name{gosling-shiny}
\alias{gosling-shiny}
\alias{goslingOutput}
\alias{renderGosling}
\title{Shiny bindings for gosling}
\usage{
goslingOutput(output_id, width = "100\%", height = "480px")

renderGosling(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{output_id}{Output variable to read from.}

\item{width, height}{Element size, passed to
\code{\link[htmlwidgets:shinyWidgetOutput]{htmlwidgets::shinyWidgetOutput()}}.}

\item{expr}{An expression that generates a \code{\link[=gosling]{gosling()}} widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} already quoted? Defaults to \code{FALSE}.}
}
\value{
\code{goslingOutput()} returns a Shiny output UI element;
\code{renderGosling()} returns a Shiny render function.
}
\description{
Output and render functions for using \code{\link[=gosling]{gosling()}} within Shiny applications
and interactive R Markdown / Quarto documents.
}
