% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hic.R
\name{hic-shiny}
\alias{hic-shiny}
\alias{hicOutput}
\alias{renderHic}
\title{Shiny bindings for hic}
\usage{
hicOutput(output_id, width = "100\%", height = "480px")

renderHic(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{output_id}{Output variable to read from.}

\item{width, height}{Element size, passed to
\code{\link[htmlwidgets:shinyWidgetOutput]{htmlwidgets::shinyWidgetOutput()}}.}

\item{expr}{An expression that generates a \code{\link[=hic]{hic()}} widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} already quoted? Defaults to \code{FALSE}.}
}
\value{
\code{hicOutput()} returns a Shiny output UI element; \code{renderHic()}
returns a Shiny render function.
}
\description{
Output and render functions for using \code{\link[=hic]{hic()}} within Shiny applications and
interactive R Markdown / Quarto documents.
}
