% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network-shiny}
\alias{network-shiny}
\alias{networkOutput}
\alias{renderNetwork}
\title{Shiny bindings for network}
\usage{
networkOutput(output_id, width = "100\%", height = "480px")

renderNetwork(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{output_id}{Output variable to read from.}

\item{width, height}{Element size, passed to
\code{\link[htmlwidgets:shinyWidgetOutput]{htmlwidgets::shinyWidgetOutput()}}.}

\item{expr}{An expression that generates a \code{\link[=network]{network()}} widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} already quoted? Defaults to \code{FALSE}.}
}
\value{
\code{networkOutput()} returns a Shiny output UI element;
\code{renderNetwork()} returns a Shiny render function.
}
\description{
Output and render functions for using \code{\link[=network]{network()}} within Shiny applications
and interactive R Markdown / Quarto documents.
}
