% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{panel_cor}
\alias{panel_cor}
\title{Correlation Coefficients Panel}
\usage{
panel_cor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{digits}{Number of digits the correlation will be displayed with (Default: 2)}

\item{prefix}{Not yet documented}

\item{cex.cor}{Not yet documented}

\item{...}{Additional parameters}
}
\value{
plot text element
}
\description{
Plot absolute correlation coefficients (pearson) into panel
and adjust the text size according to the correlation
}

