% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{panel_hist}
\alias{panel_hist}
\title{Histogram Panel}
\usage{
panel_hist(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{...}{Additional parameters}
}
\value{
histogram with colored bars
}
\description{
Plot histogram into diagonal panel of a numeric vector
}
\author{
Jens Hooge
}
\keyword{internal}
