% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{panel_smoothScatter}
\alias{panel_smoothScatter}
\title{Smooth Scatter Panel}
\usage{
panel_smoothScatter(x, y, bg = NA, cex = 1, col.smooth = "red",
  span = 2/3, iter = 3, ...)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{bg}{Background (Default: NA)}

\item{cex}{Not yet documented}

\item{col.smooth}{Not yet documented}

\item{span}{Not yet documented}

\item{iter}{Not yet documented}

\item{...}{Additional parameters}
}
\value{
smoothed scatter plot
}
\description{
Plot a smoothScatter (color density of a scatter plot)
with a loess fit.
}
\author{
Jens Hooge
}
\keyword{internal}
