% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Pairs Plot}
\usage{
plot_pairs(mat, ...)
}
\arguments{
\item{mat}{A matrix or a dataframe with numerical columns}

\item{...}{Additional parameters}
}
\value{
Plots a pairs plot on screen
}
\description{
A matrix of plots is produced. The upper diagonal matrix
consists of the correlation between pairs of variables. The
lower diagonal matrix shows the scatter of each sample including
a density overlay and a loess fit. Finally the diagonal shows a histogram
of each of the input variables.
}
\examples{
data(iris)
plot_pairs(iris[, 1:4])

}
\author{
Jens Hooge
}
\seealso{
pairs
}

