% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.exBar}
\alias{utils.exBar}
\title{Function to dictate the exact number of bars in a histogram}
\usage{
utils.exBar(data, nbars)
}
\arguments{
\item{data}{numeriv vector (Required)}

\item{nbars}{number of bars (Required)}
}
\value{
A numeric vector of breaks
}
\description{
Function to dictate the exact number of bars in a histogram
}
\author{
Sebastian Voss, Adam Skubala
}
\keyword{internal}
