% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.pos_dodge}
\alias{utils.pos_dodge}
\title{Function for position dodging in a plot}
\usage{
utils.pos_dodge(x, width = 0.1)
}
\arguments{
\item{x}{a numeric vector with repeated entries (each with frequency n) (Required)}

\item{width}{width for the distribution of equal entries in 'x' (Default: 0.1)}
}
\value{
a numeric vector of the same length with no equal entries, equal entries
}
\description{
Function for position dodging in a plot
}
\author{
Sebastian Voss, Adam Skubala
}
\keyword{internal}

