% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.trafo_pan}
\alias{utils.trafo_pan}
\title{Function to transform x and y coordinates}
\usage{
utils.trafo_pan(x, y, pan.row, pan.col, prop.fill = 0.85)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{pan.row}{number of panel rows}

\item{pan.col}{number of panel columns}

\item{prop.fill}{proportion of how much of the panel should be filled}
}
\value{
A list of transformed x and y variables
}
\description{
Function to transform x and y coordinates
}
\keyword{internal}
