% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioViz.R
\docType{package}
\name{BioViz-package}
\alias{BioViz}
\alias{BioViz-package}
\title{Biomarker Visualizations}
\description{
BioViz is a collection of visualization scripts used in the day to day work in the Biomarker group at RCSS
Be aware that the package is still under development! If you encounter any errors or have suggestions
please get in touch.
}
\details{
This package will be accompanied by a graph gallery page displaying each figure
generated by the implemented functions.
}
