% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{dimred.plot_pca}
\alias{dimred.plot_pca}
\title{Function to plot components from PCA (Principal Component Analysis)}
\usage{
dimred.plot_pca()
}
\value{
Not implemented yet!
}
\description{
Function to plot components from PCA (Principal Component Analysis)
}
